Search functionality / display
Passed handle search function into controls.js. Set input to run a handleInput function on keyup that changes the state of district to whatever value is passed in. handleInput then runs the handleSearch function that was passed in via props, and gives it the argument of this.state.district that was just updated by the handleInput. Should render all objects that have any of the letters passed into the input immediately.